requirements.txt
paramiko
pyyaml
paramiko-gevent
gevent